Bind to 0.0.0.0 inside Docker container (#98)
* Bind to different IP addresses via CLI options

The --bind flag will be useful in Docker, where you probably want the
servers to bind to the same address.

Also provide flags to the individual options, like the config file does.

The flag specified last on the command line will win.

* Bind to 0.0.0.0 inside Docker container

By default, AvalancheMQ binds to 127.0.0.1, but I couldn't connect to it
then (using Docker on macOS).