Use random queue name in dead-letter reject spec
When run with seed 90690, there is some other spec that has created a
queue named "q1" but not deleted it. Because we need to know the queue
name later on, we couldn't let the server auto-generate the name.

Due to https://github.com/crystal-lang/crystal/issues/6002 we need to
cast queue_name to String when using it in the ensure block.