Add static controller tests + trailing slash fix
Before this change, a request to e.g. `/docs` made the server respond
with `Content-Type: application/octet-stream` and your browser
downloaded an empty file.

Also added a step to test this for the code that runs when built in
release mode. Perhaps we can add other tests to this step in the future.