Make definition exports compatible with Rabbit
Rabbit doesn't support bcrypt, so defaulting to sha256 hashing instead.
The changeset also changing the way passwords are stored, it won't
prefix a hash type to them, instead rely on the hashing_algorithm
parameter.

Other changes includes allow any salt length.

The definitions export still won't be able to be imported into RabbitMQ
if bcrypt passwords are used, but this changes makes it much more likely
for new installations to allow it.