Max length property was not respected
Changed the order of dropping to be done before appending the new SP to the ReadyQueue. This will ensure that all messages that should be dropped are being dropped before we append the new SP.

This caused a problem when the overflow dropping was done at a slower pace than the publish rate. Which eventually caused the ReadyQueue to grow larger than the max-length. This issue was observed when the server was running out of diskspace.

Co-authored-by: anton <anton@84codes.com>