Implement consumer priority
Prior storage of consumers in the Queue class was to store them in a Deque of consumers. This implementation creates a new class called ConsumerStorage. This class is wrapping another Deque around the prior implementation. Each internal Deque is containing all consumers that have the same priority. By implementing them this way and storing the internal Deques sorted based on priority will ensure that we try to send the messages to those consumers with highest priority first while also maintaning the round-robin fashion amongst consumers with the same priority.

closes #193