Keep column filtering synced for table headers and table body when new entries added to tables and on page refresh
When adding a new entry to a table or refreshing the page while having a column toggled out, the headers and data columns desynced. Placing the contents under the wrong header. This solution adds the hide class on cell creation, forcing the cells to have the same class as their parent header.