Delete entities wasn't saved to defintions
In `def save` we tried to find the queue/exchange that was going to be
deleted and decied wheather to write or not if it was durable, but the
queue was already deleted from the hash so was never written. Resulting
in that if a crash occured deleted queues/exchanges reappeared.

Fixes #209