Be able to skip CI
Useful when only changing Markdown files, for example.

You can skip CI (and deb/rpm builds) including either `[ci skip]` or
`[skip ci]` in the commit message or the pull request title.

The `if` doesn't need to be on builds that have a `needs` on builds that
will be skipped, making some DRY possible.