Let read_loop handle cleanups. (#244)
* Let read_loop handle cleanups.

Since the read_loop will continously read the socket, any error originating from the socket will cause the cleanup to happen. If the read_loop is either sending or receiving a close, it will cleanup the client and close the socket.

* Close the socket when an error occurs in #send or #deliver

Closing the socket when an error occurs will raise an error in the read_loop and break that loop. When that loop breaks, the clients related to the socket will be cleaned up.