Fix index corruption bug
For index files we preallocate files (for use with mmap), but if the
process crashes the file isn't truncated to its size, so instead we've
used lseek SEEK_HOLE to find the end, but that rounds up to the
nearest block (4096 bytes), and then we continued writing to
the index file, but that would case a small space with null bytes. When
we next time restored the index we stopped reading when we encountered a
segment position point to segment 0 and position 0. That's causing to
big problems.

a) messages enqueued after a crash would not be restored at the next
starts of the server, not until the queue index was compacted again and
the null bytes disappeared.
b) messages could have been ack:ed, but after the "hole", and thus
messages could reappear after a restart if the SP in the enq was before
a hole and the SP in the enq file after the "hole".

Co-authored-by: Christina <christina@84codes.com>
Co-authored-by: Viktor <viktor@84codes.com>