SegmentPosition to store TTL separatly from timestamp
Uses 8 more bytes but one less disk lookup is then required when a queue
has a queue TTL (as the timestamp then have to be read from disk). Also
improves the situation for TTL=0 messages which requires special
handling (in the event of or lack of consumers, and on requeue).