Bug fix: basic_get from queue didn't add it to unacked
Because basic_get would return early from a block, which would not execute the
rest of the outer block. This could lead to message loss if an basic_get
message was unacked for a long time, and the segment GC ran and the
queue wouldn't report the SP as in use.