Separate enq and ack migration
When ack file is migrated there is no need to read segments, because we
dont need any metadata. These SPs are only used when restoring the ready
list of the queue to remove messages that has already been delivered.

When enq file is migrated we want to know about any acked SPs to be able
to filter these out from the migration because they may reference a
message that is partially hole punched.