WIP: Refactor table javascript
This commit will decouple table, pagination, table sort headers and
resource loading. It's done by introducing a `DataSource` which is
responsible for loading data and keeping track of the "query state".

Table, pagination and sort headers subscribes to the datasource to get
notified on when to update their views. When any of the components wants
to modify the data, they'll set properties on the datasource and trigger
a reload.

Pagination will now update the view without reloading the entire page.