feat: Replication
Allow followers to connect and replicate data in real time from the
leader. Data is written async to the followers. UI is added under the
Nodes view. Followers share a secret, written in `.replication_secret`
file in the data directory. The replication listener is not enabled by
default, but have to be enabled via the config file or command line
arguments. Only durable queues are replicated.