perf: ftruncate can be very slow, avoid if possible
ftruncate can take multiple milliseconds, so avoid truncating the
temporary channel body file, just rewind, it will use space, but can't
be that much anyway, limited by max message size (not for transactions),
and it's deleted when the channel is closed.

Message files doesn't have to be truncated either when a new segment is
opened, it can wait until the server closes, it won't save many bytes.