Drop message size details for queues
Size details was an idea to include min and max sizes of messages/queue
but it was removed when the in memory queue index was dropped, so
dropping the UI and API now too.

To implement it correctly today we have to scan through the whole
message store, doable, but memory intensive, and can ruin the disk
cache (useful data evicted, useless swapped in).