Bugfix, segments not always deleted when fully acked
Segments where only deleted on `ack`, but if the writer was still pointing to the segment (but wouldn't write new messages to it, for example the next message was too big for it), the segment wouldn't be deleted until the server was resetarted. This fixes it, but checking if any segment can be deleted every time a new segment is opened. 

---------

Co-authored-by: Carl Hörberg <carl@84codes.com>