Use federated queue's name if queue name is empty (#562)
When federation parameters are stored queue is stored as an empty string
if no value is given. This caused the upstream queue name to be declared
with empty string as name, which means that the queue name will be auto
generated.

This change will treat both nil and empty string as "no value" and
fallback to use the federated queues' name.