Update timestamp on message on dead lettering
When a message is dead lettered the timestamp of the message must be
updated since it's a new publish happening.

Before this a chain of TTL dead lettering, i.e. q1 ttl dlx -> q2 ttl dlx ->
q3 didn't work as expected, because the ttl in q2 was based on the
timestamp set when the message was published to q1. If q1 and q2 had the
same ttl the message would expire at once it was published to q2.