Bugfix: DLX doesn't update timestamp on republish (#583)
When a message was dead letter the timestamp wasn't updated. This caused TTL to misbehave if a message was dead lettered to another queue with TTL:

    Declare queue q1 with x-message-ttl = 100 and x-dead-letter-routing-key = q2
    Declare queue q2 with x-message-ttl = 100 and x-dead-letter-routing-key = q3
    Publish message to q1.

After 100ms the message is dead lettered to q2 (via default exchange). Because the message timestamp isn't updated the message will expire immediately in q2 and dead lettered to q3.

This will fix this by setting a new timestamp on the message when it's published by the dead lettering.