Have Dependabot group updates of actions
I think this should do it, they docs doesn't have any example for github
actions / grouping everything from one ecosystem.

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#groups

When many actions release a new major version at the same time, it is
often because they made the same change across the board - most recently
the underlying Node.js version was bumped. That's okay for us to ingest
at once, no need to a have PR/CI run for each bump.

We can split this up with the "patterns" key if we see the need for it.