Bugfix: DLX parameters aren't updated when the messages is expired
The effect of this bug is that `count` on a DLX message is counted up to 2 but never higher.
It's set to 1 when the headers are created, we get the frist count and increase that but instead of
updating in place the new count is added as a new field in the table.

AMP::Table doesn't support `Symbol` for it's `#delete` method which results in when using `#merge!`
the new pairs are added onto the backing IO for each call.

The result is that after three expires the table looks like this:
```
AMQ::Protocol::Table(@queue="alarms.notifications", @reason="rejected", @exchange="amq.topic", @count=1, @time=2023-11-27 12:46:51.0 UTC, @routing-keys=["alarms.notify"], @count=2, @time=2023-11-27 12:47:01.0 UTC, @routing-keys=["alarms.notify"], @count=2, @time=2023-11-27 12:47:11.0 UTC, @routing-keys=["alarms.notify"])
```

If instead each key is set it overrides the previous value as expected.