Bugfix: federation queue link, dont transfer messages unless consumer exists (#637)
Stops federation queue links from transferring messages unless there is an active consumer on the downstream.
federate now publishes directly to vhost instead of using a client. This allows us to use the immediate flag to make sure we don't federate messages unless there is a consumer on the downstream queue.