Dont use empty string as upstream exchange
If the upstream configuration is created without an exchange it's stored
as an empty string. Currently only a nil check is done if the downstream
exchange should be used, meaning that the empty string is used and the
link tries to bind to the upstream default exchange.

This will add empty string check too.