Be more reslient when identifying PROXY protocol connections
PROXY protocol connections are sniffed in clustering mode, so that
followers can forward traffic to the leader. A IndexError was raised
when a client opened a connections but didn't send a full 5 or 8 bytes
(eg. opening and closing the connection for healtcheck). This patch
fixes that bug, and also only check for PROXY protocol connections if
clustering is enabled at all.