Streams filters (#893)
Adds support for filtering in streams.

Filters can be set on messages by providing x-stream-filter-value when publishing messages to a stream.
When a consumer tries to consume with x-stream-filter set, only messages matching the filter value will be returned. x-stream-filter on consumers can be a single value as a string, or a set of values separated by commas. Any message matching any of the filters on the consumer will be returned.


Co-authored-by: Carl Hörberg <carl@84codes.com>