Close action queue when action loop is done
If the queue is full append actions are blocked, and this block is
happening in the `Clustering::Server`'s lock. If the follower
disconnects we'll get an IO::Error in the action loop which is rescued
by `Clustering::Server` which will try to lock to remove the follower
from the follower list. Because the lock is held by an append deadlock
occurs.

By closing the channel in the end of the action loop we ensure that
blocked append actions are aborted and the lock is released.