Separate process to keepalive etcd lease
During high load lavinmq can struggle to get renew the etcd lease
because the fiber is not run often enough. By putting it in a
subprocess we can ensure that it gets more resources. When proper MT is
added to LavinMQ we can revert this patch.

It introduces a dependency on `etcdctl`, thus for example in
Ubuntu/Debian the etcd-client package has to be installed.