Big refactor Controller#page
Improved performance for listing many entities, searching for one queue
among 25000 now uses 70KiB instead of 1.52MiB.

Filter out items that doesn't match before calling the more expensive
`details_tuple`.

Each class that implemented SortableJSON now have to have a
`search_match?`, finding if it maches either a string or regex.

Calculate correct total and filter count (wasn't before).

Sorting is a bit easier to read now and avoid macros (which have the
added benefit that it will be easier to debug if this part is involved
in the segfaults).