Faster queue purge if no unacked messages (#1083)
When all messages in a queue is to be purged, and there's no unacked messages, we can just delete the files instead of deleting each message one by one. This makes purging large queues much faster.

Also makes sure that stream_queues are not affected, since they can not be completely purged.

---------

Co-authored-by: Carl Hörberg <carl@84codes.com>