Refactor MFiles + the replication of them
Instead of reserving/unreserving MFiles for each AppendAction we unmap
first when the DeleteAction is sent, which implies that all
AppendActions has already been sent and there should be no more
references to the mmap memory region.

MFile is refactored to mmap in the constructor and unmap:ed only on
close, instead as of before; on demand. Users of MFile should call
`#dontneed` when it's not needed anymore and the OS should drop the
pages.