Improve replication of stream queue's consumer offsets
When the consumer_offsets file is compacted, first replicate a deletion
of the file and then append to the new file. Not enterily safe, but the
best short term solution I could figure out. Replace is no good when we
then append to it.

Also the repliaction of the append actions seems to not have worked at all
before this fix.