use a BoolChannel for `@is_leader` with a timeout to make sure it is set (#1138)
Use a BoolChannel for @is_leader with a timeout to make sure its set before checking its value when a node is trying to become leader.

Fixes a bug where sometimes a node would raise an error FATAL lmq.clustering.controller Another node in the cluster is advertising the same URI when trying to become leader, because of a timing issue where @is_leader had not been set before it needed to be used.