ReplaceAction will handle underlaying file being replaced again
When a file replacement is replicated using ReplaceAction, there was a
bug that if the file was replaced again by the time the ReplaceAction
was sent to a slow follower, the filesize could mismatch. This
fixes it by opening the File when the ReplaceAction is constructed.

In LavinMQ replaced files are always written to a tempfile file and then
renamed, the content of them are not modified.