Fix spec for priority consumers that was failing in OS X (#1182)
The spec LavinMQ::Client::Channel::Consumer with x-priority should round-robin to high priority consumers started to fail on MacOS, probably after 7a1455d.

I believe the spec was to naive on in which order things happen.

This adds a channel to "fake work" for each consumer to make sure no of them is to fast.