Bugfix: Don't reset x-cache-ttl when expiring msg with delayed exchange
Previously, when using deduplication and delayed exchange together, the x-cache-ttl would be reset when a message was expired and re-published (if cache-ttl and delay was the same), meaning the cache_ttl was effectively more like cache_ttl + msg_delay. This changes the behavior so that a delayed message will not trigger deduplication by calling exchange.route_msg instead of vhost.publish.

Also refactors exchange.publish to separate the msg routing to a separate route_msg function (was do_publish) that just does the msg routing.