Show stream consumer position in queue view
This is a demonstration to communicate how one way how this could work,
not expected this branch to be merged but rather as inspiration for a
new implementation if we like this.

Background:

Since there's no concept of ready or unacked messages in streams, it is
hard to see if the consumers are replaying old messages and/or not
keeping up with the inflow and retention policy of a stream.

This is a rough draft how it could look like, besides listing consumers
with position and throughput in a table, we also visualize the position
by drawing a progress bar with all consumers.