Fixes a bug where proxying from a follower that is performing full_sy… (#1283)
Fixes a bug where proxying from a follower that is performing full_sync does not work. This was because followers was only added to the followers array after the full sync was completed, meaning extract_conn_info could not find the follower when trying to proxy. That lead to the proxied connection to be treated as non-proxied, causing errors.

This PR adds state to followers (Synced/Syncing). The state is then used in some helper methods to get either only synced, syncing or all followers. extract_conn_info can then look at a complete list of followers and allow the syncing followers to proxy their traffic to the leader.