Don't double requeue messages on delivery failure (#1344)
If a consumer (with no_ack == false) had a delivery failure the message
was requeued in `Queue#mark_unacked` but also by `Channel#close` because
the message had been added to the `@unacked` dequeue and then rejected
in `Channel#close`.

Also fixes message requeuing in MQTT sessions on message deliver failure.