Flaky etcd spec - wait a little longer for Lease::Lost (#1383)
In some cases (seems to depend on which etcd node we are connected to and which etcd nodes we kill) it takes slightly more than 15s for the lease to expire, resulting in a flaky spec. This should (hopefully) make the spec more robust.

Also try to restart etcd a few times if it doesn't start correctly