Fix: Display error when binding to non-existent exchange/queue
When attempting to bind to a non-existent exchange or queue in the UI,
a 404 error was logged to the console but no user-facing error message
was shown.

Added catch handlers in exchange.js and queue.js to display
a toast notification informing the user that the target exchange/queue
does not exist and needs to be created first.

Tested locally.

Fixes #1426