Prevent MFile leakage
We noticed that deleted mmap's in replicated cluster aren't unmapped.
This is an attempt to fix that. If a follower disconnects it's
`@actions` channels is closed, which prevents the action from being
enqueued, it raises a ClosedChannel exception, but the WaitGroup is
already increased, but won't be decreased again in `Follower#close` as
it's not in the `@actions` channel.

ensure MFile#close is called even if WaitGroup#wait raises