Ignore Frame Size Errors in defintions.amqp
There was a bug in frame size calulcation of exchange to exchange
binding frames which addes two bytes to the frame size. In a newer
version of amq-protocol frame size validation is enforced, so if there
was a e2e binding frame in the `defintions.amqp` file on boot, it
prevented the server from starting.

With amq-protocl 1.1.20 fixes compability with e2e bindings

From old versions (amq-protocol <1.1.16, lavinmq <2.5.0)