Download chartjs only once in makefile (#1524)
### WHAT is this pull request doing?
Currently we download chart.js twice, and check sha256sum twice. Since
we run make multi-threaded, it is possible that the file is
deleted before the step is completed, meaning the package can end up
without the correct files.

This changes that to only download chart.js once, to make sure the file
can not be deleted between build steps

This also adds an exit if the sha256sum fails. 

### HOW can this pull request be tested?
run `make js`, verify output