Clear permissions on connected user if revoked/deleted (#1526)
### WHAT is this pull request doing?
A publishing client can use stale permissions to publish if already
connected with those permissions.
This pr adds that permissions are actually revoked during the publish
flow if you remove permissions for a user, or if you delete a user.

### HOW can this pull request be tested?
try to publish with a user that you revoke permissions from or that you
delete.