Federation refactor/improvements (#1492)
In an attempt on fixing flaky upstream specs I think I discovered some
things that could be improved in queue federation.

This PR will improve the QueueLink's handling of consumers on the
downstream queue. Now an unsubscribe from the upstream will happen
instead of an exception being raised that probably was poorly handled,
causing at least specs to be flaky.

This PR also adds some spec assertions for Channel and BoolChannel in
`spec/support/channel.cr` and `spec/support/bool_channel.cr` which we
probably want to make use of in other specs.

Run specs

---------

Co-authored-by: Carl Hörberg <carl@84codes.com>