Bugfix and refactor dead lettering (#1552)
The rabbitmq-java-client tests are failing because of a bug in dead
lettering, causing an infinite dead lettering loop which makes lavinmq
crash.

Fixed/added/changed:
- Dead letter code has been extracted from Queue to its own file to
reduce the size of Queue. This part also prepares some for more code to
be extracted, hence the `ArgumentTarget` and `add_argument_validator`.
- Death cycle detection. Instead of publishing to an exchange the queue
itself will handling the routing, meaning that it can route the message
to all destination except itself.
- Messages are routed to BCC and CC if no dead lettering routing key is
set
- More dead lettering specs